career_growth_assistant/requirements.txt
langchain==1.0.3
langchain-groq==1.0.0
langchain-core==1.0.4
langchain-community==0.4.1
langgraph==0.1.3

# Groq LLM
groq==0.33.0

# Data & APIs
requests==2.32.5
pandas==2.1.0
numpy==2.3.4

# JSON processing
jsonschema==4.25.1

# Utilities
pydantic==2.12.4
pyyaml==6.0.3

# Optional: for advanced features
aiohttp==3.13.2

# Development
python-dotenv==1.2.1
